requirements.txt
fastapi
uvicorn
openai
chromadb
pydantic
requests
python-dotenv
tiktoken